requirements.txt
Django==4.0.3
djangorestframework==3.13.1
Markdown==3.3.6
django-filter==21.1
django-cors-headers==3.11.0
drf-writable-nested==0.6.3